change update in backend and fix some frontend thingies

+++ frontend/src/pages/Delete.js
@@ -6,6 +6,7 @@ import {Link, Redirect} from "react-router-dom";
 import Button from "../components/Button";
 import {useState} from "react";
 import NavBar from "../components/NavBar";
+import Error from "../components/Error";
 
 
 export default function Delete() {
@@ -13,6 +14,9 @@ export default function Delete() {
     const [deletedUser, setDeletedUser] = useState()
 
     const handleDelete = () => {
+        if(user.role === "admin"){
+            return <Error>Nope</Error>
+        }
         deleteUser(user.userName)
             .then(deletedUser => setDeletedUser(deletedUser))
             .then(logout)
